@@ -18,14 +18,14 @@
     "mdlint": "remark . -f",
     "publish:dry-run": "npm run build:prod && cp -f package.json dist && cp -f README.md dist && cp -f LICENSE dist && cd dist && npm publish --dry-run",
     "publish:npm": "npm run build:prod && cp -f package.json dist && cp -f README.md dist && cp -f LICENSE dist && cd dist && npm publish",
-    "test": "jest --detectOpenHandles --forceExit",
-    "test:watch": "jest --watchAll --coverage --detectOpenHandles --forceExit",
-    "test:watch:func": "npm run test:watch -- func-test",
-    "test:coverage": "jest --coverage",
-    "test:coverage-only-changed": "jest --coverage --onlyChanged",
+    "test": "PERCY_LOGLEVEL=silent jest --detectOpenHandles --forceExit",
+    "test:watch": "PERCY_LOGLEVEL=silent jest --watchAll --coverage --detectOpenHandles --forceExit",
+    "test:watch:func": "PERCY_LOGLEVEL=silent npm run test:watch -- func-test",
+    "test:coverage": "PERCY_LOGLEVEL=silent jest --coverage",
+    "test:coverage-only-changed": "PERCY_LOGLEVEL=silent jest --coverage --onlyChanged",
     "test:ci": "npx percy exec -- npx jest --verbose --coverage --runInBand --detectOpenHandles --forceExit",
-    "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand --detectOpenHandles --forceExit",
-    "test:updateSnapshot": "jest --updateSnapshot",
+    "test:debug": "PERCY_LOGLEVEL=silent node --inspect-brk node_modules/.bin/jest --runInBand --detectOpenHandles --forceExit",
+    "test:updateSnapshot": "PERCY_LOGLEVEL=silent jest --updateSnapshot",
     "test:vulnerabilities": "snyk test",
     "ncu": "npx ncu",
     "typecheck": "npx tsc --p tsconfig.json"
@@ -41,18 +41,18 @@
   "license": "Apache-2.0",
   "homepage": "https://design.infor.com",
   "devDependencies": {
-    "@babel/core": "^7.13.10",
-    "@babel/parser": "^7.13.12",
+    "@babel/core": "^7.13.14",
+    "@babel/parser": "^7.13.13",
     "@babel/plugin-proposal-class-properties": "^7.13.0",
     "@babel/plugin-proposal-decorators": "^7.13.5",
     "@babel/plugin-transform-runtime": "^7.13.10",
     "@babel/preset-env": "^7.13.12",
-    "@babel/traverse": "^7.13.0",
-    "@babel/types": "^7.13.12",
-    "@percy/cli": "^1.0.0-beta.42",
+    "@babel/traverse": "^7.13.13",
+    "@babel/types": "^7.13.14",
+    "@percy/cli": "^1.0.0-beta.44",
     "@percy/puppeteer": "^2.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.20.0",
     "@wordpress/jest-puppeteer-axe": "^3.0.1",
     "autoprefixer": "^10.2.5",
     "babel-eslint": "^10.1.0",
@@ -60,9 +60,9 @@
     "babel-plugin-add-module-exports": "^1.0.4",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^8.1.0",
-    "css-loader": "^5.1.4",
+    "css-loader": "^5.2.0",
     "cssnano": "^4.1.10",
-    "eslint": "^7.22.0",
+    "eslint": "^7.23.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-compat": "3.9.0",
     "eslint-plugin-import": "^2.22.1",
@@ -81,7 +81,7 @@
     "jest": "^26.6.3",
     "jest-puppeteer": "^4.4.0",
     "loglevel": "^1.7.1",
-    "mini-css-extract-plugin": "^1.3.9",
+    "mini-css-extract-plugin": "^1.4.0",
     "nodemon": "^2.0.7",
     "pino": "^6.11.2",
     "postcss-loader": "^5.2.0",
@@ -100,7 +100,7 @@
     "terser-webpack-plugin": "^5.1.1",
     "typescript": "^4.2.3",
     "webpack": "^5.28.0",
-    "webpack-cli": "^4.5.0",
+    "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2"
   }
 }
