@@ -1,8 +1,13 @@
 import {
   IdsElement,
   customElement,
-  scss
+  scss,
+  mix,
+  props
 } from '../ids-base/ids-element';
+
+import { IdsEventsMixin } from '../ids-base/ids-events-mixin';
+import { IdsThemeMixin } from '../ids-base/ids-theme-mixin';
 
 // @ts-ignore
 import styles from './ids-separator.scss';
@@ -11,12 +16,23 @@ import styles from './ids-separator.scss';
  * IDS Separator Component
  * @type {IdsSeparator}
  * @inherits IdsElement
+ * @mixes IdsEventsMixin
+ * @mixes IdsThemeMixin
+ * @part separator - the menu separator element
  */
 @customElement('ids-separator')
 @scss(styles)
-class IdsSeparator extends IdsElement {
+class IdsSeparator extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin) {
   constructor() {
     super();
+  }
+
+  connectedCallback() {
+    super.connectedCallback();
+  }
+
+  static get properties() {
+    return [props.MODE, props.VERSION];
   }
 
   template() {
@@ -24,7 +40,7 @@ class IdsSeparator extends IdsElement {
     if (this.parentElement?.tagName === 'IDS-MENU-GROUP') {
       tagName = 'li';
     }
-    return `<${tagName} class="ids-separator"></${tagName}>`;
+    return `<${tagName} part="separator" class="ids-separator"></${tagName}>`;
   }
 }
 
